requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
PyPDF2==3.0.1
pyzbar==0.1.9
typing_extensions==4.12.0
